+++ src/types/enum.ts
@@ -26,6 +26,12 @@ export enum ResourceType {
   Product = 'product',
 }
 
+export enum ResourceTypeAlias {
+  Collection = 'collections',
+  Dataset = 'datasets',
+  Product = 'maps',
+}
+
 export enum ResourceStatus {
   draft = 'underDevelopment',
   complete = 'completed',